Add test checking for comment parse error
