+++ tests/parse/fail/consecutive_comments.pj
@@ -1,0 +1,3 @@
+# this is a comment
+# this is another comment
+foo_bar
