+++ tests/parse/pass/bin_op.pj
