+++ tests/parse/pass/bit_op.pj
