+++ tests/parse/pass/cmp_and_shift.pj
