+++ tests/parse/pass/let_bind.pj
