+++ tests/parse/pass/literal.pj
