+++ tests/parse/pass/logic_op.pj
