+++ tests/parse/pass/single_comment.pj
